+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ArduinoJson: change log
 =======================
 
-HEAD
-----
+v7.0.4 (2024-03-12)
+------
 
 * Make `JSON_STRING_SIZE(N)` return `N+1` to fix third-party code (issue #2054)